Add endpoint to recieve and log user actions
When a user clicks a button on the slack message, we receive a POST
callback at `/slack/actions`. In this case, parse the incoming JSON and
determine if it is either a vote or the click to end voting.

For now, log this to STDOUT. This will then later trigger an according
update to the original Slack message.